requirements.txt
cffi==2.0.0
cryptography==46.0.3
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pytest==9.0.2
